+++ scripts/classes/dice-stats-main.js
@@ -172,41 +172,41 @@ class DiceStatsTracker {
      * Method used that parses messages from the chat. This is how we know a roll has happened, what die was rolled, and the value
      * @param {Message} msg - chat message object
      */
-    parseMessage(msg){
-        let isBlind = msg.blind;
-
-        //Get Associated player object
+    parseMessage(msg)
+    {
+        // Get the player that the roll is associated with
         let playerInfo = this.PLAYER_DATA_MAP.get(msg.user.id);
 
-        //For multiple rolls in chat
-        for (let tempRoll = 0; tempRoll < msg.rolls.length; tempRoll++) {
-            
-            //For multiple dice types per roll
-            for(let tempDie=0; tempDie<msg.rolls[tempRoll]?.dice.length ; tempDie++){
+        // Get the specific system parser to parse msg
+        let parser = MESSAGE_PARSER_FACTORY.createMessageParser();
 
-                //Get die type
-                let sides = msg.rolls[tempRoll]?.dice[tempDie].faces;
-                let dieType = DS_GLOBALS.MAX_TO_DIE.get(sides);
-                let newNumbers = [];
-                //Get type of roll (Atack, Save, ect)
-                let rollType = DICE_STATS_UTILS.getRollType(msg);
+        // Parse the msg (Parser returns ROLL_OBJ[] which has some Hit / Miss stats + DIE_OBJ[]  
+        let rollInfoAry = parser.parseMsgRoll(msg);
+        // Parser Should now get deleted here as we dont need it anymore once we have the ary
+        //delete parser;
 
-                //In case there's more than one die rolled in a single instance as in 
-                //  fortune/misfortune rolls or multiple hit dice save each roll
-                newNumbers = msg.rolls[tempRoll].dice[tempDie].results.map(result => result.result)
+        // Save Each ROLL_INFO  from array into players local data
+        // TODO: Update player & Die Stats to take in {DS_ROLL_INFO} object
+        let updatedLocalRollValue = false;
+        for(let rollIT=0; rollIT<rollInfoAry.length; rollIT++){
+            let dieAry = rollInfoAry[rollIT].DiceInfo;
 
-                newNumbers.forEach(element => {
-                    playerInfo.saveRoll(isBlind, element, dieType, rollType)
-                });
+            for(let dieIT=0; dieIT<dieAry.length; dieIT++){
+                let dieInfo = dieAry[dieIT];
+
+                playerInfo.saveRoll(dieInfo?.IsBlind, dieInfo?.RollValue, 
+                                    dieInfo?.DieType, dieInfo?.RollType);
+
+                updatedLocalRollValue = true;
             }
-            
         }
 
-        //If AutoSave is Enabled by GM
+        //If AutoSave is Enabled by GM, only save updates to YOUR ROLLS to the DB
+        //  Each person updates their own DB values but loads everyones in on joining the game
         if(game.settings.get(DS_GLOBALS.MODULE_ID, DS_GLOBALS.MODULE_SETTINGS.ENABLE_AUTO_DB)) 
         {
             //If it was my roll save my data to the db
-            if(msg.user.id == game.user.id)
+            if(msg.user.id == game.user.id && updatedLocalRollValue)
             {
                 this.saveMyPlayerData();
             }