+++ scripts/classes/forms/dice-stats-tabedplayerstatspage.js
@@ -315,10 +315,5 @@ class CustomTabFormClass extends FormApplication
     activateListeners(html) {
         super.activateListeners(html);
         html.on('click', "[data-action]", this._handleButtonClick.bind(this));
-    }
-
-    async _updateObject(event, formData) {
-        return;
-    }
-    
+    }    
 }